+++ CHANGES.rst
@@ -38,6 +38,9 @@ Bug Fixes
 
 - ``astropy.table``
 
+  - Fixed reference cycle in tables that could prevent ``Table`` objects
+    from being freed from memory. [#2879]
+
   - Fixed an issue where ``Table.pprint()`` did not print the header to
     ``stdout`` when ``stdout`` is redirected (say, to a file). [#2878]